+++ rust/operator-binary/src/crd/affinity.rs
@@ -43,14 +43,13 @@ mod tests {
           name: simple-nifi
         spec:
           image:
-            productVersion: 1.27.0
+            productVersion: 2.7.2
           clusterConfig:
             authentication:
               - authenticationClass: nifi-admin-credentials-simple
             sensitiveProperties:
               keySecret: simple-nifi-sensitive-property-key
               autoGenerate: true
-            zookeeperConfigMapName: simple-nifi-znode
           nodes:
             roleGroups:
               default: