+++ rust/operator-binary/src/main.rs
@@ -3,9 +3,10 @@
 #![allow(clippy::result_large_err)]
 use std::sync::Arc;
 
+use anyhow::anyhow;
 use clap::Parser;
 use crd::v1alpha1::NifiClusteringBackend;
-use futures::{FutureExt, StreamExt};
+use futures::{FutureExt, StreamExt, TryFutureExt};
 use stackable_operator::{
     YamlSchema,
     cli::{Command, RunArguments},
@@ -16,7 +17,7 @@ use stackable_operator::{
         core::v1::{ConfigMap, Service},
     },
     kube::{
-        ResourceExt,
+        CustomResourceExt as _, ResourceExt,
         core::DeserializeGuard,
         runtime::{
             Controller,
@@ -28,11 +29,13 @@ use stackable_operator::{
     logging::controller::report_controller_reconciled,
     shared::yaml::SerializeOptions,
     telemetry::Tracing,
+    utils::signal::{self, SignalWatcher},
 };
 
 use crate::{
     controller::NIFI_FULL_CONTROLLER_NAME,
     crd::{NifiCluster, NifiClusterVersion, authorization::NifiOpaConfig, v1alpha1},
+    webhooks::conversion::create_webhook_server,
 };
 
 mod config;
@@ -44,12 +47,14 @@ mod product_logging;
 mod reporting_task;
 mod security;
 mod service;
+mod webhooks;
 
 mod built_info {
     include!(concat!(env!("OUT_DIR"), "/built.rs"));
 }
 
 const OPERATOR_NAME: &str = "nifi.stackable.tech";
+const FIELD_MANAGER: &str = "nifi-operator";
 
 #[derive(Parser)]
 #[clap(about, author)]
@@ -65,7 +70,7 @@ async fn main() -> anyhow::Result<()> {
         Command::Crd => NifiCluster::merged_crd(NifiClusterVersion::V1Alpha1)?
             .print_yaml_schema(built_info::PKG_VERSION, SerializeOptions::default())?,
         Command::Run(RunArguments {
-            operator_environment: _,
+            operator_environment,
             watch_namespace,
             product_config,
             maintenance,
@@ -88,21 +93,36 @@ async fn main() -> anyhow::Result<()> {
                 description = built_info::PKG_DESCRIPTION
             );
 
+            // Watches for the SIGTERM signal and sends a signal to all receivers, which gracefully
+            // shuts down all concurrent tasks below (EoS checker, controller).
+            let sigterm_watcher = SignalWatcher::sigterm()?;
+
             let eos_checker =
                 EndOfSupportChecker::new(built_info::BUILT_TIME_UTC, maintenance.end_of_support)?
-                    .run()
+                    .run(sigterm_watcher.handle())
                     .map(anyhow::Ok);
-
-            let product_config = product_config.load(&[
-                "deploy/config-spec/properties.yaml",
-                "/etc/stackable/nifi-operator/config-spec/properties.yaml",
-            ])?;
 
             let client = stackable_operator::client::initialize_operator(
                 Some(OPERATOR_NAME.to_string()),
                 &common.cluster_info,
             )
             .await?;
+
+            let webhook_server = create_webhook_server(
+                &operator_environment,
+                maintenance.disable_crd_maintenance,
+                client.as_kube_client(),
+            )
+            .await?;
+
+            let webhook_server = webhook_server
+                .run(sigterm_watcher.handle())
+                .map_err(|err| anyhow!(err).context("failed to run webhook server"));
+
+            let product_config = product_config.load(&[
+                "deploy/config-spec/properties.yaml",
+                "/etc/stackable/nifi-operator/config-spec/properties.yaml",
+            ])?;
 
             let event_recorder = Arc::new(Recorder::new(
                 client.as_kube_client(),
@@ -133,7 +153,6 @@ async fn main() -> anyhow::Result<()> {
                     watch_namespace.get_api::<ConfigMap>(&client),
                     watcher::Config::default(),
                 )
-                .shutdown_on_signal()
                 .watches(
                     client
                         .get_api::<DeserializeGuard<auth_core::v1alpha1::AuthenticationClass>>(&()),
@@ -156,6 +175,7 @@ async fn main() -> anyhow::Result<()> {
                             .map(|nifi| ObjectRef::from_obj(&*nifi))
                     },
                 )
+                .graceful_shutdown_on(sigterm_watcher.handle())
                 .run(
                     controller::reconcile_nifi,
                     controller::error_policy,
@@ -183,7 +203,12 @@ async fn main() -> anyhow::Result<()> {
                 )
                 .map(anyhow::Ok);
 
-            futures::try_join!(nifi_controller, eos_checker)?;
+            let delayed_nifi_controller = async {
+                signal::crd_established(&client, v1alpha1::NifiCluster::crd_name(), None).await?;
+                nifi_controller.await
+            };
+
+            futures::try_join!(delayed_nifi_controller, eos_checker, webhook_server)?;
         }
     }